Add secondary objective for optimization
For local minima, it often occurs that neighbors have the same objective
(makespan or tardiness). We add a second objective to further
distinguish between possible selections.

For now, we use the average objective value of the selection's neighbors
as the secondary objective. Since all neighbors are evaluated in the
intensification phase anyway, this does not introduce any computational
overhead.